stack yamls: use crds, move commentary into separate file
Two main changes here. The commentary has been moved into an
example.stack.yaml file so that it isn't cluttering up the stack.yaml
files in the sample stacks. Also, since we want to use CRDs for
configuring the template bindings, so that we have built-in schema
validation, this changeset also configures the hooks in a more
CRD-specific way.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>